AI-PLANT-DISEASE/requirements.txt
tensorflow>=2.10.0
keras>=2.10.0
flask>=2.3.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
opencv-python>=4.7.0
Pillow>=9.3.0
scikit-learn>=1.2.0
/